Bug fixes
Fixed a bug where the datepicker would populate a date when the input
box was cleared and any key was pressed. Also fixed a bug on the check
in page where if the filter was applied (on either plant or inventory)
then upon submission any selected items hidden by the filter were not
included.